Don’t force rerender in htmlbars tests
We need to force a rerender (similar to `scheduleRevalidation`) for
glimmer because we don’t have runloop integration yet, but we intend to
remove that soon. However, we should not force the same for htmlbars so
we don’t accidentally regress functionality.

Also, rename `inZone` to `runTask`.
