+++ packages/ember-htmlbars/tests/utils/abstract-test-case.js
@@ -1,0 +1,1 @@
+../../../ember-glimmer/tests/utils/abstract-test-case.js
